+++ changelog.md
@@ -35,3 +35,4 @@
  * Fix alloy blast recipe autogeneration with CraftTweaker materials
  * New structure for the Rotary Hearth Furnace
  * Reworked Parallel for the Rotary Hearth Furnace and Bulk Blast Chiller
+ * Deprecated Chemical Plant multiblock for implementation in another mod. It will be REMOVED in the following update.